Initiate an ATO

This is a wrapper around the usual way of creating S4 objects through new(). The main purpose of this function is to ensure the ATO is created with the right table type. See table_type for more details.

Usage

init_ato(det, dep, tag, ani, obs)

Arguments

det

an object of class ATO_det. See make_det.

dep

an object of class ATO_dep. See make_dep.

tag

an object of class ATO_tag. See make_tag.

ani

an object of class ATO_ani. See make_ani.

obs

an object of class ATO_obs. See make_obs.

Value

an ATO