Make an ATO animals object

Formats the input data into the ATO format and appends the ATO_ani class.

Usage

make_ani(
  animal = NA_character_,
  capture_location = NA_character_,
  capture_datetime = as.POSIXct(NA_real_),
  capture_lat = NA_real_,
  capture_lon = NA_real_,
  release_location = NA_character_,
  release_datetime = as.POSIXct(NA_real_),
  release_lat = NA_real_,
  release_lon = NA_real_,
  tz,
  ...
)

Arguments

animal

Name of the animal that received the tag, character.

capture_location

Name of the location where the animal was captured, character.

capture_datetime

date and time of the capture, posixct.

capture_lat

latitude of the capture. Preferably in WGS84, numeric.

capture_lon

longitude of the capture. Preferably in WGS84, numeric.

release_location

Name of the location where the animal was released, character.

release_datetime

date and time of the release, posixct.

release_lat

latitude of the release. Preferably in WGS84, numeric.

release_lon

longitude of the release. Preferably in WGS84, numeric.

tz

the timezone of the datetime data.

...

Non-standard columns to be added to the table.

Value

an ATO_ani object, ready to be used by add or init_ato.

See also