Make an ATO deployments object

Formats the input data into the ATO format and appends the ATO_dep class.

Usage

make_dep(
  receiver_model = NA_character_,
  receiver_serial = NA_integer_,
  receiver_codeset = NA_character_,
  deploy_location = NA_character_,
  deploy_datetime = as.POSIXct(NA_real_),
  deploy_lat = NA_real_,
  deploy_lon = NA_real_,
  deploy_z = NA_real_,
  recover_datetime = as.POSIXct(NA_real_),
  recover_lat = NA_real_,
  recover_lon = NA_real_,
  transmitter = NA_character_,
  transmitter_manufacturer = NA_character_,
  transmitter_ping_rate = NA_real_,
  transmitter_model = NA_character_,
  transmitter_serial = NA_integer_,
  tz,
  ...
)

Arguments

receiver_model

Model of the receiver, character.

receiver_serial

Receiver serial number, integer.

receiver_codeset

Codeset of the receiver, character.

deploy_location

Name of the location where the receiver was deployed, character.

deploy_datetime

date and time of the deployment, posixct.

deploy_lat

latitude of the deployment. Preferably in WGS84, numeric.

deploy_lon

longitude of the deployment. Preferably in WGS84, numeric.

deploy_z

depth of the deployment, as measured from the reference surface of the water body, numeric.

recover_datetime

date and time of the recovery, posixct.

recover_lat

latitude of the recovery point. Preferably in WGS84, numeric.

recover_lon

longitude of the recovery point. Preferably in WGS84, numeric.

transmitter

Transmitter code for a beacon/reference tag, character.

transmitter_manufacturer

Manufacturer of the transmitter, character.

transmitter_ping_rate

Expected ping rate of the transmitter, numeric. In seconds.

transmitter_model

Model of the transmitter, character.

transmitter_serial

Serial number of the transmitter, integer.

tz

the timezone of the datetime data.

...

Non-standard columns to be added to the table.

Value

an ATO_dep object, ready to be used by add or init_ato.

See also